+++ SDK/m90app/xpos-sdk-demo/src/vas/electricity.h
@@ -41,7 +41,7 @@ protected:
     std::string phoneNumber;
     PaymentMethod payMethod;
     EnergyType energyType;
-    long long amount;
+    unsigned long amount;
 
     CardPurchase cardPurchase;
