+++ SDK/m90app/xpos-sdk-demo/src/vas/smile.h
@@ -30,7 +30,7 @@ protected:
     std::string customerID;
     std::string phoneNumber;
     PaymentMethod payMethod;
-    long long amount;
+    unsigned long amount;
 
     CardPurchase cardPurchase;
